requirements.txt
python-dotenv==0.7.1
Flask==0.12.2
mypy==0.550
